+++ osu.Server.Queues.ScorePump/Performance/PerformanceMain.cs
@@ -3,14 +3,12 @@
 
 using McMaster.Extensions.CommandLineUtils;
 using osu.Server.Queues.ScorePump.Performance.Totals;
+using osu.Server.Queues.ScorePump.Performance.Values;
 
 namespace osu.Server.Queues.ScorePump.Performance
 {
-    [Command(Name = "performance", Description = "Computes the performance (pp) of scores.")]
-    [Subcommand(typeof(AllCommand))]
-    [Subcommand(typeof(SqlCommand))]
-    [Subcommand(typeof(UsersCommand))]
-    [Subcommand(typeof(ScoresCommand))]
+    [Command(Name = "update-pp", Description = "Computes the performance (pp) of scores.")]
+    [Subcommand(typeof(UpdateValuesMain))]
     [Subcommand(typeof(UpdateTotalsMain))]
     public sealed class PerformanceMain
     {
