+++ osu.Server.Queues.ScorePump/Performance/Values/AllCommand.cs
@@ -11,7 +11,7 @@ using Dapper;
 using McMaster.Extensions.CommandLineUtils;
 using osu.Server.Queues.ScoreStatisticsProcessor;
 
-namespace osu.Server.Queues.ScorePump.Performance
+namespace osu.Server.Queues.ScorePump.Performance.Values
 {
     [Command(Name = "all", Description = "Computes pp of all users.")]
     public class AllCommand : PerformanceCommand
